backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.3
langgraph==0.2.59
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.13
opik>=1.9.0
PyGithub==2.5.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.20
